+++ docs/changes.rst
@@ -40,6 +40,8 @@ v0.11.2 (UNRELEASED)
 
 - Add :attr:`mopidy.models.Track.disc_no` field. (Partly fixes: :issue:`286`)
 
+- Add :attr:`mopidy.models.Album.num_discs` field. (Partly fixes: :issue:`286`)
+
 
 v0.11.1 (2012-12-24)
 ====================